requirements.txt
flask==3.1.0
python-dotenv==1.0.1
firecrawl-py==1.6.3
pydantic==2.10.3
